LPS-56219 Rename package
