+++ modules/apps/export-import/export-import-test/test/integration/com/liferay/exportimport/PortletDataContextReferencesTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.lar;
+package com.liferay.exportimport;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.bookmarks.model.BookmarksEntry;
