+++ modules/apps/export-import/export-import-test/test/integration/com/liferay/exportimport/dependencies/layout_references.txt
